Move permission policy tests to a submodule
