+++ wagtail/tests/permission_policies/test_page_permission_policies.py
@@ -4,7 +4,9 @@ from django.test import TestCase
 from wagtail.models import GroupPagePermission, Page, get_default_page_content_type
 from wagtail.permission_policies.pages import PagePermissionPolicy
 from wagtail.test.utils import WagtailTestUtils
-from wagtail.tests.test_permission_policies import PermissionPolicyTestUtils
+from wagtail.tests.permission_policies.test_permission_policies import (
+    PermissionPolicyTestUtils,
+)
 
 
 class PermissionPolicyTestCase(PermissionPolicyTestUtils, WagtailTestUtils, TestCase):
